scripts/requirements.txt
# scripts/requirements.txt

google-generativeai
requests
beautifulsoup4
PyMuPDF
PyGithub
google-generativeai
